% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_splitDS.R
\name{create_data_splitDS}
\alias{create_data_splitDS}
\title{Separate Training and Test data}
\usage{
create_data_splitDS(
  data_name,
  data_classes,
  output_var,
  drop_columns,
  train_test_ratio
)
}
\arguments{
\item{data_name}{Name of the data.}

\item{data_classes}{List of data class for all columns.}

\item{output_var}{The name of the column containing the output.}

\item{drop_columns}{Which columns should be removed from the data.}

\item{train_test_ratio}{Ratio of training data of the whole data.}
}
\value{
The training-test split.
}
\description{
Separate Training and Test data
}
