% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format_checkDS.R
\name{data_format_checkDS}
\alias{data_format_checkDS}
\title{Checking the Data for unwanted behaviour.}
\usage{
data_format_checkDS(
  data_name,
  bounds_and_levels,
  output_var,
  loss_function,
  drop_NA
)
}
\arguments{
\item{data_name}{Name of the data.}

\item{bounds_and_levels}{List of all columns which stores either bounds or
level sets for all variables.}

\item{output_var}{The name of the output variable.}

\item{loss_function}{The loss function which we use to optimize our boosted
tree.}

\item{drop_NA}{If NA-values should be deleted.}
}
\value{
The data classes for all columns.
}
\description{
Checking the Data for unwanted behaviour.
}
