% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_spp_candDS.R
\name{gen_spp_candDS}
\alias{gen_spp_candDS}
\title{Title Create splitting-point candidates.}
\usage{
gen_spp_candDS(data_name, cand_select = NULL)
}
\arguments{
\item{data_name}{Name of the data.}

\item{cand_select}{By which method the candidates shall be selected.}
}
\value{
The splitting-point candidates.
}
\description{
Title Create splitting-point candidates.
}
