% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_trainingDS.R
\name{save_trainingDS}
\alias{save_trainingDS}
\title{Split up data in training features and training output}
\usage{
save_trainingDS(data_name)
}
\arguments{
\item{data_name}{Name of the data.}
}
\value{
The split.
}
\description{
Split up data in training features and training output
}
