% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_treeDS.R
\name{save_treeDS}
\alias{save_treeDS}
\title{Save the prediction through a tree.}
\usage{
save_treeDS(data_name, tree, amt_drops)
}
\arguments{
\item{data_name}{Name of the data.}

\item{tree}{The trained tree which shall be saved.}

\item{amt_drops}{How many trees got dropped.}
}
\value{
The prediction through the newly trained tree.
}
\description{
Save the prediction through a tree.
}
