% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_variableDS.R
\name{save_variableDS}
\alias{save_variableDS}
\title{Save a variable}
\usage{
save_variableDS(data_name, variable)
}
\arguments{
\item{data_name}{Name of the data.}

\item{variable}{The variable which shall be saved.}
}
\value{
The variable which shall be saved.
}
\description{
Save a variable
}
