% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_binsDS.R
\name{split_binsDS}
\alias{split_binsDS}
\title{Split up the data into bins by slitting points.}
\usage{
split_binsDS(data_name, min_max, spp_cand, current_tree, data_type)
}
\arguments{
\item{data_name}{The name of the data.}

\item{min_max}{The maximum and minimum values of the features.}

\item{spp_cand}{The Splitting-point candidates.}

\item{current_tree}{The currently trained tree.}

\item{data_type}{The type of data per feature.}
}
\value{
The histogram bins for each split in all features.
}
\description{
Split up the data into bins by slitting points.
}
