% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_treeDS.R
\name{train_treeDS}
\alias{train_treeDS}
\title{Train a remote tree.}
\usage{
train_treeDS(data_name)
}
\arguments{
\item{data_name}{Name of the data.}
}
\value{
The trained tree.
}
\description{
Train a remote tree.
}
