% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hist_initDS.R
\name{calc_hist_initDS}
\alias{calc_hist_initDS}
\title{Calculate the predicted output and histograms}
\usage{
calc_hist_initDS(data_name)
}
\arguments{
\item{data_name}{Name of the data.}
}
\value{
The training features and calculated output and histograms.
}
\description{
Calculate the predicted output and histograms
}
