% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_histDS.R
\name{calc_histDS}
\alias{calc_histDS}
\title{Calculate the histograms.}
\usage{
calc_histDS(data_name, removed_trees = NULL)
}
\arguments{
\item{data_name}{Name of the data.}

\item{removed_trees}{Which trees got removed for training.}
}
\value{
The training features and calculated output and histograms.
}
\description{
Calculate the histograms.
}
