% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_splitDS.R
\name{create_data_splitDS}
\alias{create_data_splitDS}
\title{Separate Training and Test data}
\usage{
create_data_splitDS(data_name, train_test_ratio)
}
\arguments{
\item{data_name}{Name of the data.}

\item{train_test_ratio}{Ratio of training data of the whole data.}
}
\value{
The training-test split.
}
\description{
Separate Training and Test data
}
