% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitDS.R
\name{data_splitDS}
\alias{data_splitDS}
\title{Reduce the data to the last added leaves}
\usage{
data_splitDS(
  data_name,
  training_features,
  bounds_and_levels,
  current_tree,
  data_classes
)
}
\arguments{
\item{data_name}{Name of the data.}

\item{training_features}{All the training data per features.}

\item{bounds_and_levels}{List of maximum and minimum value for numeric and
levels for factor features.}

\item{current_tree}{The currently trained tree.}

\item{data_classes}{Data class for all features.}
}
\value{
The two last added leafs of the tree.
}
\description{
Reduce the data to the last added leaves
}
