% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitDS.R
\name{data_splitDS}
\alias{data_splitDS}
\title{Reduce the data to the last added leaves}
\usage{
data_splitDS(training_features, min_max, current_tree)
}
\arguments{
\item{training_features}{All the training data per features.}

\item{min_max}{List of maximum and minimum value per feature.}

\item{current_tree}{The currently trained tree.}
}
\value{
The two last added leafs of the tree.
}
\description{
Reduce the data to the last added leaves
}
