% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessiansDS.R
\name{hessiansDS}
\alias{hessiansDS}
\title{Retrieve the hessian bins}
\usage{
hessiansDS(data_name)
}
\arguments{
\item{data_name}{The name of the data.}
}
\value{
The hessian bins.
}
\description{
Retrieve the hessian bins
}
