% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessiansDS.R
\name{hessiansDS}
\alias{hessiansDS}
\title{Retrieve the hessian bins}
\usage{
hessiansDS(data_name, bounds_and_levels, spp_cand, data_classes)
}
\arguments{
\item{data_name}{The name of the data.}

\item{bounds_and_levels}{List of maximum and minimum value for numeric and
levels for factor features.}

\item{spp_cand}{The Splitting-point candidates.}

\item{data_classes}{Data class per feature.}
}
\value{
The hessian bins.
}
\description{
Retrieve the hessian bins
}
