% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveleafDS.R
\name{saveleafDS}
\alias{saveleafDS}
\title{Save the split leaves on the server.}
\usage{
saveleafDS(data_name, bounds_and_levels, data_classes, current_tree)
}
\arguments{
\item{data_name}{The name of the data.}

\item{bounds_and_levels}{List of maximum and minimum value for numeric and
levels for factor features.}

\item{data_classes}{Denotes for each feature if it is numeric or a factor.}

\item{current_tree}{The tree which gets currently trained.}
}
\value{
The newly created leaves.
}
\description{
Save the split leaves on the server.
}
