% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_binsDS.R
\name{split_binsDS}
\alias{split_binsDS}
\title{Split up the data into bins by slitting points.}
\usage{
split_binsDS(data_name)
}
\arguments{
\item{data_name}{The name of the data.}
}
\value{
The histogram bins for each split in all features.
}
\description{
Split up the data into bins by slitting points.
}
