% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_evaluationDS.R
\name{tree_evaluationDS}
\alias{tree_evaluationDS}
\title{Evaluate an output of a data-point under a tree.}
\usage{
tree_evaluationDS(data_point, tree)
}
\arguments{
\item{data_point}{The data-point to be evaluated.}

\item{tree}{The tree which maps the data.}
}
\value{
The output under the tree.
}
\description{
Evaluate an output of a data-point under a tree.
}
