% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_full_treeDS.R
\name{update_full_treeDS}
\alias{update_full_treeDS}
\title{Save a sum of prediction through single trees.}
\usage{
update_full_treeDS(data_name, removed_trees, last_tree_nmb)
}
\arguments{
\item{data_name}{Name of the data.}

\item{removed_trees}{Which trees got removed for training.}

\item{last_tree_nmb}{How many trees have been trained in total.}
}
\value{
The full predicition.
}
\description{
Save a sum of prediction through single trees.
}
