% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_treeDS.R
\name{update_treeDS}
\alias{update_treeDS}
\title{Update weight of the prediction of a tree.}
\usage{
update_treeDS(data_name, tree_number, amt_drops, amt_adds)
}
\arguments{
\item{data_name}{Name of the data.}

\item{tree_number}{Which tree shall be updated.}

\item{amt_drops}{How many trees were dropped for the training.}

\item{amt_adds}{How many trees were added this round}
}
\value{
The adapted tree.
}
\description{
Update weight of the prediction of a tree.
}
