% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_weightDS.R
\name{update_weightDS}
\alias{update_weightDS}
\title{Update weights}
\usage{
update_weightDS(data_name, current_tree)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{current_tree}{The fully chosen tree which misses its weights.}
}
\value{
The weights for each leaf.
}
\description{
Update weights
}
